Fix animation glitch

+++ source/Irrlicht/os.cpp
@@ -211,8 +211,11 @@ namespace os
 
 	u32 Timer::getRealTime()
 	{
-        double time = emscripten_get_now();
-        return (u32)(time);
+		double time = emscripten_get_now();
+		// `time` is over 1.7e+12. If converted directly to u32,
+		// it will overflow and yield 0. Cast to u64 first
+		// to get correct wrapping.
+	        return (u32)((u64)time);
 	}
 } // end namespace os
 
@@ -295,7 +298,6 @@ namespace os
 	{
 		if (isStopped())
 			return LastVirtualTime;
-
 		return LastVirtualTime + (u32)((StaticTime - StartRealTime) * VirtualTimerSpeed);
 	}
